🚚 Rename mappe for tester beregningV1 -> beregning for å følge mappestruktren til koden elleres
