+++ src/test/resources/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/beregning/varierendeAktivitetsdager/flere_aktiviteterV2.feature
